Crash url session instrumentation ALPH-2195 (#51)
* fix: bug fix, now we are sending the masked image after the pipe process finished
* fix: crash for acko in URLSessionInstrumentation.injectInNSURLClasses()
This change replaces DispatchQueue.concurrentPerform with a serial for loop to avoid concurrent modification of Objective-C runtime structures, which could lead to unsafe behavior and crashes. It introduces an NSLock and a didInject flag to ensure injectInNSURLClasses() is executed only once in a thread-safe manner. This prevents simultaneous instrumentation attempts from multiple threads during SDK initialization. The change ensures that swizzling is safe and idempotent, resolving a known crash in OpenTelemetry Swift SDK v1.9.1. The rest of the instrumentation logic remains unchanged to preserve behavior and compatibility.

* fix: remove session recording init